requirements.txt
fastapi
uvicorn[standard]
torch
pillow
